Still more IO
Now starting to store faces into a matrix, and also now needs two arguments (one for an input file name, one for an output file name) as per the specs, and writes out an output file. 

NEXT UP:
Work on storing and printing out the faces. Then work on a loop accepting arguments to perform operations on the vertices.