IO done??
Now will successfully read in and print out models the same as they come in. The only difference is that now there is a float format to the vertices inputted, regardless of if they came in as integers or not. I can ask if that's okay and deal with it if it's not since technically there is a difference between the input and output files.