model
This is a modified version of the P1 code, now set to create a Model object rather than only handle one thing at a time. All that's been changed right now is the package and class name. I'll need to go through the Main (or maybe the readFile() method?) and adapt it into an init() method, and remove all the components that require navigating a Main to do things. Remove some I/O handling in order to let method declarations and stuff handle it.